Moved Shortcut related classes to shortcut package
